refactor : directory overhauled and abstrations optimized
