+++ LMS/api/users/models.py
@@ -1,3 +1,1 @@
-from django.db import models
-from .__models__.user_model import Users
-
+from .user_models.user_model import Users
