+++ LMS/api/users/serializers.py
@@ -1,0 +1,1 @@
+from user_serializers.user_serializer import *
