+++ LMS/pyproject.toml
@@ -16,6 +16,9 @@ bcrypt = "^4.1.2"
 pillow = "^10.2.0"
 django-rest-auth = "^0.9.5"
 django-allauth = "^0.60.1"
+pydantic = "^2.5.3"
+pydantic-settings = "^2.1.0"
+python-dotenv = "^1.0.0"
 
 
 [build-system]
